+++ system/src/apps/_module.mjs
@@ -1,2 +1,5 @@
 export {default as FalloutChemDoses} from "./FalloutChemDoses.mjs";
 export {default as FalloutPartySleep} from "./FalloutPartySleep.mjs";
+export {default as ItemSelector} from "./CompendiumItemSelectors/ItemSelector.mjs";
+export {default as ItemTypeMenu} from "./ItemTypeMenu.mjs";
+export {default as SourceFilterSettings} from "./SourceFilterSettings.mjs";