+++ system/src/sheets/FalloutPcSheet.mjs
@@ -45,6 +45,11 @@ export default class FalloutPcSheet extends FalloutBaseActorSheet {
 	activateListeners(html) {
 		super.activateListeners(html);
 
+		html.find(".availability-roll").click(async event => {
+			event.preventDefault();
+			this.actor.rollAvailabilityCheck();
+		});
+
 		html.find(".power-armor-monitor-health-value").change(event => {
 			event.preventDefault();
 
@@ -53,7 +58,7 @@ export default class FalloutPcSheet extends FalloutBaseActorSheet {
 
 			let apparel = this.actor.items.get(apparelId);
 
-			if (apparel && apparel.system.appareltype === "powerArmor") {
+			if (apparel && apparel.system.apparelType === "powerArmor") {
 				apparel.update({ "system.health.value": newHealthValue });
 			}
 		});
@@ -282,7 +287,7 @@ export default class FalloutPcSheet extends FalloutBaseActorSheet {
 			const list = context.items.filter(
 				i => i.type === type
 			).filter(
-				i => i.system.appareltype === subType
+				i => i.system.apparelType === subType
 			);
 
 			return {